Refactor plugin system
